Implement dashboard, sessions, settings pages and conflict resolution
- DashboardPage: Replace hardcoded stats with real store data, add quick
  actions and recent sessions display
- SessionsPage: Full session management with create/delete/export,
  template support, and session analytics
- SettingsPage: Three-tab settings (Agent, Git, UI) with localStorage
  persistence
- ParallelExecutionPage: Wire up conflict resolution to backend
- Add conflicts_resolve command to backend for conflict resolution
- Fix git worktree creation to use WorktreeAddOptions with branch ref
- Fix markdown parser regex patterns for metadata extraction
- Fix YAML parser to reject empty/comment-only documents
- Exclude test files from tsconfig to resolve type conflicts

+++ src-tauri/src/commands/sessions.rs
@@ -4,7 +4,6 @@ use crate::database::{self, Database};
 use crate::models::{AgentType, Session, SessionConfig, SessionStatus};
 use chrono::Utc;
 use serde::{Deserialize, Serialize};
-use std::path::PathBuf;
 use tauri::State;
 use uuid::Uuid;
 
@@ -466,7 +465,7 @@ pub async fn compare_sessions(
         session2_name: session2.name,
         tasks_completed_diff: tasks1_completed - tasks2_completed,
         total_cost_diff: session1.total_cost - session2.total_cost,
-        total_tokens_diff: session1.total_tokens - session2.total_tokens,
+        total_tokens_diff: (session1.total_tokens - session2.total_tokens) as i64,
         config_differences,
         performance_summary,
     })
